requirements.txt
pyparsing
pyyaml
six
enum34
ddt
